STAR-964: Add new parameter to startClient functions
The new parameter failOnConnectionError controls the behaviour
of the startClient functions. If it is set to true and the
initial connection attempt to the server, which is done in
startClient, fails, the startClient call returns true. If it
is set to false the clientStart call returns true and the
worker thread is started which will now start to try to connect
to the server until stop() is called.