STAR-1582 Fix deadlock on client stop (#32)
Fix deadlock when shutting down a client, which would get stuck indefinitly 1 out of 5 times in the main thread when calling stop. The
deadlock was casued by the main thread getting stuck while
waiting for the client thread to join. At the same time, the client worker thread was stuck waiting for recv to return a value. In case no data was sent on the network and the remote side did not disconnect, the main thread will be stuck indefinitly. This commit solves the issue by first closing the SRT socket before trying to join the thread, instead of after, as earlier. This makes the recv call return an error and exit the thread loop.